Make problems #1-5 into modules (+ #40)
This is the first step at an attempt of making a single program able
to run all of the problems at once, in the aim of adding timing info
at some point.

There are unanswered questions for now. Does the file names starting
with lowercase also work on case sensitive file systems (ie on Linux
for example) ? (Module names must begin with a capital letter).

Do we want to actually use this file system/module mapping ? If not
using it, then we must specify all files to ghc individually (but
maybe cabal or stack can handle this for us).

Do we want to remove the 001, 002 and other intermediate folders ?
They were nice to hide away build oject files and input data
files. Note the data files will have to be moved elsewhere anyway
since the running program won't know about module paths.