Add cabal file (empty for now) to identify project root
Without this file, haskell-mode tries to regenerate a TAGS file in
every subdirectory for definitions it cannot find. The cabal file
helps identify the project root and keep a single TAGS file there.

I generate tags with: hasktags . --ignore-close-implementation -e
The ignore switch is nice to have a single point to jump to when
type signature and definition are only a few lines apart. (By default
it adds a point for each, forcing you to pick one when jumping.)