Generate stack.yaml with stack init
Note complete command was:

  stack init --install-ghc --resolver lts-9.16 --solver

First switch is to install ghc the first time using stack, which is
required to use the resolver. The resolver switch is set to lts-9.16
specifically (current version is 9.17), because the integer-gmp
package's cabal file uses the new ^> "caret" operator in 9.17, which
my version of stack cannot parse. Finally the solver switch is for
memoize, which isn't part of hackage but gets added to the extra-deps
directive.

Motivation to use stack was for use with intero, which looks pretty
awesome. Currently haskell-mode is broken with ghc 8.2.x (intero is
dependent on haskell-mode), recommended version is still 8.0.x.